add Device acquire/release test
