+++ sdl2-sonic-drivers/test/drivers/TestMIDDriver.cpp
@@ -15,52 +15,12 @@
 #include <spdlog/sinks/ostream_sink.h>
 #include <sstream>
 #include <algorithm>
+#include <drivers/midi/devices/SpyDevice.hpp>
+#include <drivers/MIDDriverMock.hpp>
 
 
 namespace drivers
 {
-    namespace midi::devices
-    {
-        class SpyDevice : public Device
-        {
-            virtual void sendEvent(const audio::midi::MIDIEvent& e) const noexcept override
-            {
-
-            };
-            virtual void sendMessage(const uint8_t msg[], const uint8_t size) const noexcept override
-            {
-
-            }
-            virtual void sendSysEx(const audio::midi::MIDIEvent& e) const noexcept override
-            {
-
-            }
-            virtual void pause() const noexcept override
-            {
-
-            }
-            virtual void resume() const noexcept override
-            {
-
-            }
-        };
-    }
-
-    class MIDDriverMock : public MIDDriver
-    {
-    public:
-        explicit MIDDriverMock(const std::shared_ptr<audio::scummvm::Mixer>& mixer, const std::shared_ptr<midi::Device>& device) :
-            MIDDriver(mixer, device)
-        {
-
-        }
-
-        void protected_processTrack(const audio::midi::MIDITrack& track, const uint16_t division)
-        {
-            processTrack(track, division);
-        }
-    };
-
     TEST(MIDDriver, SEQUENCE_NAME_META_EVENT)
     {
         using audio::stubs::StubMixer;
